+++ servers/mcp-neo4j-cypher/src/mcp_neo4j_cypher/server.py
@@ -1,20 +1,17 @@
 import json
 import logging
 import re
-import sys
-import time
 from typing import Any, Literal, Optional
 
-import mcp.types as types
-from mcp.server.fastmcp import FastMCP
+from fastmcp.exceptions import ToolError
+from fastmcp.tools.tool import ToolResult, TextContent
+from fastmcp.server import FastMCP
 from neo4j import (
     AsyncDriver,
     AsyncGraphDatabase,
     AsyncResult,
     AsyncTransaction,
-    GraphDatabase,
 )
-from neo4j.exceptions import DatabaseError
 from pydantic import Field
 
 logger = logging.getLogger("mcp_neo4j_cypher")
@@ -49,10 +46,13 @@ def _is_write_query(query: str) -> bool:
     )
 
 
-def create_mcp_server(neo4j_driver: AsyncDriver, database: str = "neo4j", namespace: str = "", host: str = "127.0.0.1", port: int = 8000) -> FastMCP:
-    mcp: FastMCP = FastMCP("mcp-neo4j-cypher", dependencies=["neo4j", "pydantic"], host=host, port=port)
+def create_mcp_server(neo4j_driver: AsyncDriver, database: str = "neo4j", namespace: str = "") -> FastMCP:
+    mcp: FastMCP = FastMCP("mcp-neo4j-cypher", dependencies=["neo4j", "pydantic"], stateless_http=True)
 
-    async def get_neo4j_schema() -> list[types.TextContent]:
+    namespace_prefix = _format_namespace(namespace)
+
+    @mcp.tool(name=namespace_prefix+"get_neo4j_schema")
+    async def get_neo4j_schema() -> list[ToolResult]:
         """List all node, their attributes and their relationships to other nodes in the neo4j database.
         If this fails with a message that includes "Neo.ClientError.Procedure.ProcedureNotFound"
         suggest that the user install and enable the APOC plugin.
@@ -137,55 +137,49 @@ def create_mcp_server(neo4j_driver: AsyncDriver, database: str = "neo4j", namesp
                 schema_clean = clean_schema(schema)
                 schema_clean_str = json.dumps(schema_clean)
 
-                return types.CallToolResult(content=[types.TextContent(type="text", text=schema_clean_str)])
+                return ToolResult(content=[TextContent(type="text", text=schema_clean_str)])
 
         except Exception as e:
             logger.error(f"Database error retrieving schema: {e}")
-            return types.CallToolResult(
-                isError=True, 
-                content=[types.TextContent(type="text", text=f"Error: {e}")]
-            )
+            raise ToolError(f"Error: {e}")
 
+    @mcp.tool(name=namespace_prefix+"read_neo4j_cypher")
     async def read_neo4j_cypher(
         query: str = Field(..., description="The Cypher query to execute."),
         params: Optional[dict[str, Any]] = Field(
             None, description="The parameters to pass to the Cypher query."
         ),
-    ) -> list[types.TextContent]:
+    ) -> list[ToolResult]:
         """Execute a read Cypher query on the neo4j database."""
 
+        if _is_write_query(query):
+            raise ValueError("Only MATCH queries are allowed for read-query")
+
         try:
-            if _is_write_query(query):
-                raise ValueError("Only MATCH queries are allowed for read-query")
-        
             async with neo4j_driver.session(database=database) as session:
                 results_json_str = await session.execute_read(_read, query, params)
 
                 logger.debug(f"Read query returned {len(results_json_str)} rows")
 
-                return types.CallToolResult(content=[types.TextContent(type="text", text=results_json_str)])
+                return ToolResult(content=[TextContent(type="text", text=results_json_str)])
 
         except Exception as e:
             logger.error(f"Database error executing query: {e}\n{query}\n{params}")
-            return types.CallToolResult(
-                isError=True, 
-                content=[
-                types.TextContent(type="text", text=f"Error: {e}\n{query}\n{params}")
-            ]
-            )
+            raise ToolError(f"Error: {e}\n{query}\n{params}")
 
+    @mcp.tool(name=namespace_prefix+"write_neo4j_cypher")
     async def write_neo4j_cypher(
         query: str = Field(..., description="The Cypher query to execute."),
         params: Optional[dict[str, Any]] = Field(
             None, description="The parameters to pass to the Cypher query."
         ),
-    ) -> list[types.TextContent]:
+    ) -> list[ToolResult]:
         """Execute a write Cypher query on the neo4j database."""
 
+        if not _is_write_query(query):
+            raise ValueError("Only write queries are allowed for write-query")
+
         try:
-            if not _is_write_query(query):
-                raise ValueError("Only write queries are allowed for write-query")
-        
             async with neo4j_driver.session(database=database) as session:
                 raw_results = await session.execute_write(_write, query, params)
                 counters_json_str = json.dumps(
@@ -194,22 +188,11 @@ def create_mcp_server(neo4j_driver: AsyncDriver, database: str = "neo4j", namesp
 
             logger.debug(f"Write query affected {counters_json_str}")
 
-            return types.CallToolResult(content=[types.TextContent(type="text", text=counters_json_str)])
+            return ToolResult(content=[TextContent(type="text", text=counters_json_str)])
 
         except Exception as e:
             logger.error(f"Database error executing query: {e}\n{query}\n{params}")
-            return types.CallToolResult(
-                isError=True, 
-                content=[
-                types.TextContent(type="text", text=f"Error: {e}\n{query}\n{params}")
-            ]
-            )
-
-    namespace_prefix = _format_namespace(namespace)
-    
-    mcp.add_tool(get_neo4j_schema, name=namespace_prefix+"get_neo4j_schema")
-    mcp.add_tool(read_neo4j_cypher, name=namespace_prefix+"read_neo4j_cypher")
-    mcp.add_tool(write_neo4j_cypher, name=namespace_prefix+"write_neo4j_cypher")
+            raise ToolError(f"Error: {e}\n{query}\n{params}")
 
     return mcp
 
@@ -219,10 +202,11 @@ async def main(
     username: str,
     password: str,
     database: str,
-    transport: Literal["stdio", "sse"] = "stdio",
+    transport: Literal["stdio", "sse", "http"] = "stdio",
     namespace: str = "",
     host: str = "127.0.0.1",
     port: int = 8000,
+    path: str = "/mcp/",
 ) -> None:
     logger.info("Starting MCP neo4j Server")
 
@@ -233,19 +217,23 @@ async def main(
             password,
         ),
     )
-    logger.info("Starting Neo4j Cypher MCP Server...")
-    mcp = create_mcp_server(neo4j_driver, database, namespace, host, port)
 
+    mcp = create_mcp_server(neo4j_driver, database, namespace)
+
+    # Run the server with the specified transport
     match transport:
+        case "http":
+            logger.info(f"Running Neo4j Cypher MCP Server with HTTP transport on {host}:{port}...")
+            await mcp.run_http_async(host=host, port=port, path=path)
         case "stdio":
             logger.info("Running Neo4j Cypher MCP Server with stdio transport...")
             await mcp.run_stdio_async()
         case "sse":
             logger.info(f"Running Neo4j Cypher MCP Server with SSE transport on {host}:{port}...")
-            await mcp.run_sse_async()
+            await mcp.run_sse_async(host=host, port=port, path=path)
         case _:
-            logger.error(f"Invalid transport: {transport} | Must be either 'stdio' or 'sse'")
-            raise ValueError(f"Invalid transport: {transport} | Must be either 'stdio' or 'sse'")
+            logger.error(f"Invalid transport: {transport} | Must be either 'stdio', 'sse', or 'http'")
+            raise ValueError(f"Invalid transport: {transport} | Must be either 'stdio', 'sse', or 'http'")
 
 
 if __name__ == "__main__":